+++ Dalti/View/NewMessages/NewMessageView.swift
@@ -7,39 +7,6 @@
 
 import SwiftUI
 import SDWebImageSwiftUI
-
-class NewMessageViewModel: ObservableObject {
-    //checked
-    @Published var users = [ChatUser]()
-    @Published var errorMessage = ""
-    
-    init() {
-        fetchAllUsers()
-    }
-    
-    private func fetchAllUsers() {
-//        FirebaseManager.shared.firestore.collection("Community")
-//            .document("Users").collection(FirebaseConstants.users)
-        FirebaseManager.shared.firestore.collection("Community")
-            .document("Users").collection("User")
-            .getDocuments { documentsSnapshot, error in
-                if let error = error {
-                    self.errorMessage = "Failed to fetch users"
-                    print("Failed to fetch users: \(error)")
-                    return
-                }
-                
-                documentsSnapshot?.documents.forEach({ snapshot in
-                    let user = try? snapshot.data(as: ChatUser.self)
-                    if user?.uid != FirebaseManager.shared.auth.currentUser?.uid {
-                        self.users.append(user!)
-                    }
-                    
-                })
-            }
-    }
-}
-
 struct NewMessageView: View {
     
     let didSelectNewUser: (ChatUser) -> ()
@@ -52,7 +19,6 @@ struct NewMessageView: View {
         NavigationStack {
             ScrollView {
                 Text(vm.errorMessage)
-                
                 ForEach(vm.users) { user in
                     Button {
                         presentationMode.wrappedValue.dismiss()
