+++ Dalti/ar.lproj/Localizable.strings
@@ -51,7 +51,7 @@
  "Forget Password" = "نسيت كلمة المرور";
  "Pleas click bellow to read the privacy policy page on the web" = "اذا كنت تريد الإطلاع على صفحتنا بالويب لقراءة سياسة الخصوصية الخاصة بنا اضغط على الزر بالأسفل للذهاب للصفحة";
 
- "the item in the photo above is missing please contact the person who post it by call or chat within the app, help people to find their items and they will help you too.." = "الغرض بالاعلى مفقود الرجاد التواصل مع الشخص الذي قام بنشره أما بالأتصال به أو بمحادثته بالرسائل الخاصة في التطبيق، ساعد الآخرين حتى يساعدونك";
+ "the item in the photo above is missing please contact the person who post it by call or chat within the app, help people to find their items and they will help you too.." = "الغرض بالاعلى مفقود الرجاء التواصل مع الشخص الذي قام بنشره أما بالأتصال به أو بمحادثته بالرسائل الخاصة من خلال التطبيق";
  "You have arrived to collect your order." = "";
  "Do you want to check in?" = "";
